+++ libs/spark/.storybook/preview.js
@@ -1,13 +1,26 @@
 import React, { Fragment } from 'react';
 import { addDecorator } from '@storybook/react';
+import isChromatic from 'chromatic/isChromatic';
 import {
   CssBaseline,
   FontFacesBaseline,
   Unstable_FontFacesBaseline,
+  withStyles,
 } from '../src';
+
+const ChromaticCssBaseline = withStyles((theme) => ({
+  '@global': {
+    '*': {
+      animationDuration: '0.001ms !important',
+      animationIterationCount: '1 !important',
+      transitionDuration: '0.001ms !important',
+    },
+  },
+}))(() => null);
 
 addDecorator((Story) => (
   <Fragment>
+    {isChromatic() ? <ChromaticCssBaseline /> : null}
     <CssBaseline />
     <FontFacesBaseline />
     <Unstable_FontFacesBaseline />